requirements.txt
Jinja2==3.1.2
MarkupSafe==2.1.3
psutil==5.9.5
py-cpuinfo==9.0.0
